+++ geometry/sprite/sprite.go
@@ -23,7 +23,7 @@ type Mesh struct {
 	Size   object.Property[vec2.T]
 	Sprite object.Property[assets.Texture]
 
-	mesh vertex.MutableMesh[vertex.Vertex, uint16]
+	mesh vertex.MutableMesh[vertex.Vertex, uint32]
 }
 
 var _ mesh.Mesh = &Mesh{}
@@ -40,7 +40,7 @@ func New(pool object.Pool, args Args) *Mesh {
 		Sprite: object.NewProperty(args.Texture),
 	})
 
-	sprite.mesh = vertex.NewTriangles[vertex.Vertex, uint16](fmt.Sprintf("sprite_%.2f_%.2f", args.Size.X, args.Size.Y), nil, nil)
+	sprite.mesh = vertex.NewTriangles[vertex.Vertex, uint32](fmt.Sprintf("sprite_%.2f_%.2f", args.Size.X, args.Size.Y), nil, nil)
 	sprite.generate()
 
 	sprite.SetTexture(texture.Diffuse, args.Texture)
@@ -59,7 +59,7 @@ func (p *Mesh) generate() {
 		vertex.New(vec3.New(-0.5*w, 0.5*h, 0), vec3.Zero, vec2.Zero, color.White),
 		vertex.New(vec3.New(0.5*w, -0.5*h, 0), vec3.Zero, vec2.Zero, color.White),
 	}
-	indices := []uint16{
+	indices := []uint32{
 		0, 1, 2,
 		0, 3, 1,
 	}
